Schedule POCO objects instead of immediately returning them
Often when we are setting up a WhenAny (and by proxy, a Bind or OneWayBind
binding) in a constructor, some of the controls are null (depending on the UI
framework, GTK# is the current offender), meaning that the binding is broken.

This defers collecting the value until the next time the UI thread has a
chance to do something, which usually solves this problem.