Protect against Subscription exceptions
This commit protects ReactiveCommand and ReactiveAsyncCommand from
exceptions thrown inside a Subscribe, and marshals them to ThrownExceptions.

Normally Rx doesn't guard against this kind of operation and regards it as
undefined behavior, but in the ReactiveUI case, users are often doing
non-trivial operations in the Subscribe, so make their life a little saner.