OMG IT ACTUALLY WORKS
This method that @onovotny wrote, doesn't deadlock *and* doesn't fail store
cert, unlike the loaded module list trick.

The reason that previous incarnations of this didn't work, is that await +
static constructor + blocking wait in Task = deadlock. The solution is to
move the code doing the awaiting into its own class