added PS script to ReactiveUI-Core\tools for scaffolding The init.ps1 should be called on install and subsequent solution loads by NuGet Package Manager in Visual Studio, per their documentation. This script loads a PowerShell script module that sets up the generator command (Generate-RxUIObjects) and tab expansion for the template type argument of the same.
The generator searches subdirectories for the RxUIModelGenerator
executable and errors if it cannot be found.

A small change was necessary to the Program.cs of the CLI generator
as I could not find a way to prevent PowerShell from sending the
name of the invoked executable as the first argument to the exe
itself.  I'm not a PS expert, so there may be a way.  In any event
the change required was minimal; hopefully it passes muster.

All testing was done manually.  I did not create a nuget package
and install it to ensure that the init was called as their docs
state, but I did invoke it from the Package Manager Console and
ensure that the generator worked as expected.