Fix Reactive*ViewSource problems with ReloadData().
If a data source changes too many times, it is possible to break
a UI{Table,Collection}View.  Please see the docs of
lastCallToNumberOfSections introduced by this commit itself,
which fixes this problem.

This commit also adds hooks to Changing observables, thus releasing
bindings earlier than before.  I do not think that this is necessary
for correctness, but it does not seem to harm and may be beneficial in
avoiding needless work in some scenarios.