feat: upgrade reactiveui-events for Xamarin.Forms from v2.2.0.31 to v2.3.1.114 (#1150)
This also resolves a build error as the EventsGenerator will download the latest version of Xamarin Forms from NuGet and codegen eventhandlers from that. Now, if a new major/minor is released the packages.config needs to be manually updated to reflect this otherwise the build may fail until a commit like this one is done if new eventhandlers are surfaced in the public api of the new release of Xamarin.Forms.