Fix ActOnEveryObject missing changes when SuppressChangeNotifications is used (#1688)
*  fix: fixes ActOnEveryObject missing changes when the list changes are suppressed

Previously ActOnEveryObject would not handle list changes correctly if the changes were made while SuppressChangeNotifications() was used.
With this commit, ActOnEveryObject works as expected.

Fixes issue #1675.
BREAKING CHANGE: protected virtual method publishResetNotification()  is removed. It is not used anymore and current subclasses might be overriding it because it affects Reset() and SuppressChangeNotifications(), which it doesn't anymore .

* Update ApiApprovalTests.ReactiveUI.approved.txt

* Changed the approved reactiveui document