Fix #2382 DelayChangeNotifications does not delay PropertyChanged events (#2403)
* test: test deferral of changing notifications

Improve the test of deferring change notifications on reactive objects to test that "Changing" notifications are delayed in addition to "Changed" notifications. Also improved to verify which properties change notifications were sent.

* fix: make DelayChangeNotifications() apply to PropertyChanging and PropertyChanged events
DelayChangeNotifications() delayed notifications through the Changing and Changed observables, but had no effect on the PropertyChanging and PropertyChanged events. This fixes that.

* style: follow event naming convention and write some proper docs comments

Co-authored-by: Jeff Walker <jeff@deepcellbio.com>